Adds outputs & determines alarm status.
Commit adds new string outputs to the MathOperations.cs.  This will be parsed by the ReminderPage in order to how the user what the problem is.  The string output currenlty contains the answer and the problem.  Also adds the ReminderPage and begins the logic to determine if the Alarm is going off.  Has not been tested yet.